fixed problem with circular dependencies
