+++ backend/src/app/core/filters/types.py
@@ -1,0 +1,3 @@
+from typing import List, Union
+
+FilterValue = Union[bool, str, int, List[str], List[List[str]]]
